Add tags to Person model, admin (#1720)
